Revert "Authentification Implementation"
